chore: refactor from slice to actions and reducer
