+++ packages/suite/src/actions/wallet/__fixtures__/blockchainActions.ts
@@ -1,7 +1,7 @@
 import { NOTIFICATION } from '@suite-actions/constants';
 import { TRANSACTION, BLOCKCHAIN } from '@wallet-actions/constants';
 import { analyzeTransactionsFixtures as analyzeTransactions } from '@suite-common/wallet-utils';
-import { accountActions } from '@suite-common/wallet-core';
+import { accountsActions } from '@suite-common/wallet-core';
 
 const DEFAULT_ACCOUNT = {
     deviceState: 'deviceState',
@@ -38,7 +38,7 @@ const analyzeTransactionsExtended = [
         result: [BLOCKCHAIN.SYNCED],
     },
     {
-        result: [TRANSACTION.ADD, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.ADD, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [
                 { blockHeight: 5, blockHash: '5', txid: '5' },
@@ -50,7 +50,7 @@ const analyzeTransactionsExtended = [
         },
     },
     {
-        result: [TRANSACTION.ADD, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.ADD, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [{ blockHeight: undefined, blockHash: '1', txid: '1' }],
         },
@@ -62,19 +62,19 @@ const analyzeTransactionsExtended = [
         },
     },
     {
-        result: [TRANSACTION.REMOVE, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.REMOVE, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [],
         },
     },
     {
-        result: [TRANSACTION.REMOVE, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.REMOVE, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [{ blockHeight: 1, blockHash: '1', txid: '1' }],
         },
     },
     {
-        result: [TRANSACTION.REMOVE, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.REMOVE, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [],
         },
@@ -83,7 +83,7 @@ const analyzeTransactionsExtended = [
         result: [
             TRANSACTION.REMOVE,
             TRANSACTION.ADD,
-            accountActions.updateAccount.type,
+            accountsActions.updateAccount.type,
             BLOCKCHAIN.SYNCED,
         ],
         resultTxs: {
@@ -94,7 +94,7 @@ const analyzeTransactionsExtended = [
         result: [
             TRANSACTION.REMOVE,
             TRANSACTION.ADD,
-            accountActions.updateAccount.type,
+            accountsActions.updateAccount.type,
             BLOCKCHAIN.SYNCED,
         ],
         resultTxs: {
@@ -102,7 +102,7 @@ const analyzeTransactionsExtended = [
         },
     },
     {
-        result: [TRANSACTION.ADD, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.ADD, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [
                 { blockHeight: undefined, blockHash: '4', txid: '4' },
@@ -113,7 +113,7 @@ const analyzeTransactionsExtended = [
         },
     },
     {
-        result: [TRANSACTION.ADD, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.ADD, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [
                 { blockHeight: 2, blockHash: '2', txid: '2' },
@@ -123,7 +123,7 @@ const analyzeTransactionsExtended = [
         },
     },
     {
-        result: [TRANSACTION.ADD, accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+        result: [TRANSACTION.ADD, accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         resultTxs: {
             'xpub-btc-deviceState': [
                 { blockHeight: 4, blockHash: '4', txid: '4' },
@@ -140,7 +140,7 @@ const analyzeTransactionsExtended = [
         result: [
             TRANSACTION.REMOVE,
             TRANSACTION.ADD,
-            accountActions.updateAccount.type,
+            accountsActions.updateAccount.type,
             BLOCKCHAIN.SYNCED,
         ],
         resultTxs: {
@@ -155,7 +155,7 @@ const analyzeTransactionsExtended = [
         result: [
             TRANSACTION.REMOVE,
             TRANSACTION.ADD,
-            accountActions.updateAccount.type,
+            accountsActions.updateAccount.type,
             BLOCKCHAIN.SYNCED,
         ],
         resultTxs: {
@@ -173,7 +173,7 @@ const analyzeTransactionsExtended = [
         result: [
             TRANSACTION.REMOVE,
             TRANSACTION.ADD,
-            accountActions.updateAccount.type,
+            accountsActions.updateAccount.type,
             BLOCKCHAIN.SYNCED,
         ],
         resultTxs: {
@@ -226,7 +226,7 @@ export const onBlock = analyzeTransactions
             state: {
                 accounts: [{ ...DEFAULT_ACCOUNT, history: { total: 0, unconfirmed: 0 } }],
             },
-            result: [accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+            result: [accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         },
         {
             description: 'Account specific fields changed, blockbook: total',
@@ -240,7 +240,7 @@ export const onBlock = analyzeTransactions
             state: {
                 accounts: [{ ...DEFAULT_ACCOUNT, history: { total: 0, unconfirmed: 0 } }],
             },
-            result: [accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+            result: [accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         },
         {
             description: 'Account specific fields changed, ripple: sequence',
@@ -254,7 +254,7 @@ export const onBlock = analyzeTransactions
             state: {
                 accounts: [{ ...DEFAULT_ACCOUNT, networkType: 'ripple', misc: { sequence: 0 } }],
             },
-            result: [accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+            result: [accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         },
         {
             description: 'Account specific fields changed, ripple: balance',
@@ -276,7 +276,7 @@ export const onBlock = analyzeTransactions
                     },
                 ],
             },
-            result: [accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+            result: [accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         },
         {
             description: 'Account specific fields changed, ethereum: nonce',
@@ -290,7 +290,7 @@ export const onBlock = analyzeTransactions
             state: {
                 accounts: [{ ...DEFAULT_ACCOUNT, networkType: 'ethereum', misc: { nonce: 0 } }],
             },
-            result: [accountActions.updateAccount.type, BLOCKCHAIN.SYNCED],
+            result: [accountsActions.updateAccount.type, BLOCKCHAIN.SYNCED],
         },
         {
             description: 'Account does not exists',
