+++ packages/suite/src/actions/wallet/tokenActions.ts
@@ -3,11 +3,11 @@ import { Account } from '@wallet-types';
 import { Dispatch } from '@suite-types';
 import * as accountUtils from '@suite-common/wallet-utils';
 import * as notificationActions from '@suite-actions/notificationActions';
-import { accountActions } from '@suite-common/wallet-core';
+import { accountsActions } from '@suite-common/wallet-core';
 
 export const addToken = (account: Account, tokenInfo: TokenInfo[]) => (dispatch: Dispatch) => {
     dispatch(
-        accountActions.updateAccount({
+        accountsActions.updateAccount({
             ...account,
             tokens: (account.tokens || []).concat(accountUtils.enhanceTokens(tokenInfo)),
         }),
