+++ packages/suite/src/middlewares/suite/sentryMiddleware.ts
@@ -19,7 +19,7 @@ import { AppState, Action, Dispatch } from '@suite-types';
 import { ACCOUNT, BLOCKCHAIN, DISCOVERY } from '@wallet-actions/constants';
 
 import { Severity } from '@sentry/types';
-import { accountActions } from '@suite-common/wallet-core';
+import { accountsActions } from '@suite-common/wallet-core';
 
 const deviceContextName = 'trezor-device';
 
@@ -42,8 +42,8 @@ const breadcrumbActions = [
     SUITE.AUTH_DEVICE,
     DEVICE.CONNECT,
     DEVICE.DISCONNECT,
-    accountActions.createAccount.type,
-    accountActions.updateAccount.type,
+    accountsActions.createAccount.type,
+    accountsActions.updateAccount.type,
     DISCOVERY.COMPLETE,
     SUITE.UPDATE_SELECTED_DEVICE,
     SUITE.REMEMBER_DEVICE,
