+++ packages/suite/src/middlewares/wallet/walletMiddleware.ts
@@ -6,7 +6,7 @@ import { TRANSACTION, BLOCKCHAIN } from '@wallet-actions/constants';
 import * as selectedAccountActions from '@wallet-actions/selectedAccountActions';
 import * as sendFormActions from '@wallet-actions/sendFormActions';
 import * as modalActions from '@suite-actions/modalActions';
-import { accountActions } from '@suite-common/wallet-core';
+import { accountsActions } from '@suite-common/wallet-core';
 import * as receiveActions from '@wallet-actions/receiveActions';
 import * as cardanoStakingActions from '@wallet-actions/cardanoStakingActions';
 import * as coinmarketBuyActions from '@wallet-actions/coinmarketBuyActions';
@@ -25,10 +25,10 @@ const walletMiddleware =
             const accounts = api
                 .getState()
                 .wallet.accounts.filter(a => a.deviceState === deviceState);
-            api.dispatch(accountActions.removeAccount(accounts));
+            api.dispatch(accountsActions.removeAccount(accounts));
         }
 
-        if (accountActions.createAccount.match(action)) {
+        if (accountsActions.createAccount.match(action)) {
             // gather transactions from account.create action
             const account = action.payload;
             api.dispatch(transactionActions.add(account.history.transactions || [], account, 1));
@@ -43,13 +43,13 @@ const walletMiddleware =
         next(action);
 
         if (
-            accountActions.createAccount.match(action) ||
-            accountActions.updateAccount.match(action)
+            accountsActions.createAccount.match(action) ||
+            accountsActions.updateAccount.match(action)
         ) {
             api.dispatch(blockchainActions.subscribe(action.payload.symbol));
         }
 
-        if (accountActions.removeAccount.match(action)) {
+        if (accountsActions.removeAccount.match(action)) {
             api.dispatch(blockchainActions.unsubscribe(action.payload));
         }
 
@@ -89,7 +89,7 @@ const walletMiddleware =
                 (nextRouter.app === 'wallet' && nextRouter.hash !== prevRouter.hash);
         }
         if (resetReducers) {
-            api.dispatch(accountActions.disposeAccount());
+            api.dispatch(accountsActions.disposeAccount());
             api.dispatch(sendFormActions.dispose());
             api.dispatch(receiveActions.dispose());
             api.dispatch(coinmarketBuyActions.dispose());
