+++ packages/suite/src/types/wallet/index.ts
@@ -19,6 +19,7 @@ import { CardanoStakingAction } from '@wallet-actions/cardanoStakingActions';
 import { PollingAction } from '@wallet-actions/pollingActions';
 import { CoinjoinAccountAction } from '@wallet-actions/coinjoinAccountActions';
 import { CoinjoinClientAction } from '@wallet-actions/coinjoinClientActions';
+import { AccountsAction } from '@suite-common/wallet-core';
 import { NETWORKS } from '@wallet-config';
 import { ArrayElement } from '@trezor/type-utils';
 
@@ -72,4 +73,5 @@ export type WalletAction =
     | CardanoStakingAction
     | PollingAction
     | CoinjoinAccountAction
-    | CoinjoinClientAction;
+    | CoinjoinClientAction
+    | AccountsAction;
